Refactor viewCreate instead of viewManager
